T#771: Extract search routes + Meili + FTS5 to src/search/routes.ts
Phase 1.6 of Backend Modularization (Library #102). Mechanical move,
no logic changes.

**Moved to src/search/routes.ts (~493L)**:
- Meilisearch client state + initMeilisearch + backfillMeilisearch
- FTS5 virtual table creation + initial backfill
- searchUrlMap + searchUrlFor helpers
- searchIndexUpsert + searchIndexDelete (cross-domain, exported)
- sanitizeFtsQuery + fts5Search helpers
- indexSpecFiles helper
- VALID_SOURCE_TYPES constant
- 4 route handlers: GET /api/search/legacy, GET /api/search,
  POST /api/search/reindex, GET /api/search/status

**Exports**:
- registerSearchRoutes(app, sqlite, helpers) — DI for hasSessionAuth,
  isLocalNetwork, isTrustedRequest, handleSearch
- initSearch(sqliteDb) — server startup entry (synchronous FTS5 setup
  + async Meili init via fire-and-forget promise to preserve original
  initMeilisearch().then(...) shape)
- searchIndexUpsert / searchIndexDelete — for cross-domain use
  (forum, library, task, spec write handlers already import these)
- backfillMeilisearch — exposed for /api/search/reindex internally

**Server.ts changes**:
- Added import from ./search/routes.ts (line 99)
- Replaced legacy search handler block (L1922-1939, 18L) with
  one-line reference comment
- Replaced search infrastructure block (L11257-11688, 432L) with:
    initSearch(sqlite);
    registerSearchRoutes(app, sqlite, { hasSessionAuth,
      isLocalNetwork, isTrustedRequest, handleSearch });
- Removed unused 'import { MeiliSearch } from meilisearch' at L17
- Net: server.ts down 455 lines (5 insertions, 450 deletions)

**Verification**:
- Side-by-side handler diff: all 4 routes preserved byte-for-byte
  (auth gates, type-prefix syntax, id-lookup shortcuts, Meili-first/
  FTS5-fallback ordering, dedup logic)
- Boot test: server starts clean, '[MEILI] No master key configured,
  skipping' logs from new module, all endpoints listed
- tsc --noEmit: no extraction-related type errors (pre-existing
  errors only — server.ts has prior class issues unrelated to T#771)
- /api/forum/search left in server.ts (not in scope — forum domain)
- /api/help registry entries (L1736-1738, L1582) reference the routes
  by path only, no code changes needed there

T#788 follow-up (risk + prowl requireBeastIdentity migration) still
queued for next modularization-cleanup pass.
